parseInt to parseFloat to maintain decimal precision

+++ src/components/Add/Add.js
@@ -26,8 +26,8 @@ class Add extends Component {
     const tags = store.tags.filter(tag => tag.isChecked).map(tag => tag.name);
     const address = store.location.storeAddress;
     const coordinates = [
-      parseInt(store.location.storeLongitude),
-      parseInt(store.location.storeLatitude)
+      parseFloat(store.location.storeLongitude),
+      parseFloat(store.location.storeLatitude)
     ];
 
     //TODO filter/sanitize user input
